requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
openai 
pinecone-client